Ensure `remove` accepts str and Path-like objects
