+++ tests/basic_env.py
@@ -13,9 +13,9 @@ from git import Repo
 from git.exc import GitCommandNotFound
 
 from dvc.repo import Repo as DvcRepo
-from dvc.utils import remove
 from dvc.utils.compat import open
 from dvc.utils.compat import str
+from dvc.utils.fs import remove
 
 
 logger = logging.getLogger("dvc")
